+++ src/constraints/ConstrExp.cpp
@@ -886,14 +886,14 @@ void ConstrExp<SMALL, LARGE>::fixOverflow(const IntMap<int>& level, int bitOverf
   assert(bitReduce > 0);
   assert(bitOverflow >= bitReduce);
   LARGE maxVal = std::max<LARGE>(largestCoef, std::max(degree, aux::abs(rhs)) / INF);
-  if (maxVal > 0 && (int)aux::msb(maxVal) >= bitOverflow) {
+  if (maxVal > 0 && aux::msb(maxVal) >= bitOverflow) {
     assert(getCutoffVal() == maxVal);
     LARGE div = aux::ceildiv<LARGE>(maxVal, aux::powtwo<LARGE>(bitReduce) - 1);
     assert(aux::ceildiv<LARGE>(maxVal, div) <= aux::powtwo<LARGE>(bitReduce) - 1);
     weakenDivideRound(div, [&](Lit l) { return !isFalse(level, l) && l != -asserting && l != asserting; });
   } else {
     // check that largestCoef indeed is big enough
-    assert(getCutoffVal() <= 0 || (int)aux::msb(getCutoffVal()) < bitOverflow);
+    assert(getCutoffVal() <= 0 || aux::msb(getCutoffVal()) < bitOverflow);
   }
   assert(isSaturated());
   assert(hasNoZeroes());
@@ -943,7 +943,7 @@ bool ConstrExp<SMALL, LARGE>::fitsInDouble() const {
 
 template <typename SMALL, typename LARGE>
 bool ConstrExp<SMALL, LARGE>::largestCoefFitsIn(int bits) const {
-  return (int)aux::msb(getLargestCoef()) < bits;
+  return aux::msb(getLargestCoef()) < bits;
 }
 
 template <typename SMALL, typename LARGE>
